+++ packages/yourloops/test/integration/assert/signup-stepper.assert.ts
@@ -46,7 +46,7 @@ export const checkAccountSelectorStep = () => {
   expect(within(patientRadioInput).getByRole('radio')).toHaveAttribute('disabled')
   expect(within(caregiverRadioInput).getByTestId('radio-label-header')).toHaveTextContent('Caregiver and family')
   expect(within(caregiverRadioInput).getByTestId('radio-label-body')).toHaveTextContent('View patients diabetes data as an individual caregiver or family member.')
-  expect(within(hcpRadioInput).getByTestId('radio-label-header')).toHaveTextContent('Professional')
+  expect(within(hcpRadioInput).getByTestId('radio-label-header')).toHaveTextContent('Health care professional')
   expect(within(hcpRadioInput).getByTestId('radio-label-body')).toHaveTextContent('View all your patients diabetes data in one place. Join and create a care team for your clinic or practice.')
   expect(within(patientRadioInput).getByTestId('radio-label-header')).toHaveTextContent('Patient')
   expect(within(patientRadioInput).getByTestId('radio-label-body')).toHaveTextContent('If you use a DBL System, you have to create your account from your handset. You can’t create several accounts with the same email address.')