+++ packages/yourloops/test/integration/use-cases/notification-management.ts
@@ -25,39 +25,8 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import {
-  checkMonitoringAlertsIconsInactiveForFirstPatient,
-  checkPatientColumnsFiltersContent,
-  checkPatientListColumnSort,
-  checkPatientListCurrentTab,
-  checkPatientListCurrentTabForPrivateTeam,
-  checkPatientListFilters,
-  checkPatientListHeaderForHcp,
-  checkPatientListHideShowColumns,
-  checkPatientListPendingTab,
-  checkPatientListTooltipsMgDL,
-  checkPatientListTooltipsMmolL,
-  type Router
-} from '../assert/patient-list.assert'
-import { checkPatientFiltersForPrivateTeam } from '../assert/patient-filters.assert'
+import { checkAcceptNotificationDirectShare } from '../assert/notification-direct-share.assert'
 
-export const testPatientListForHcp = async (router: Router) => {
-  checkPatientListHeaderForHcp()
-  await checkPatientListTooltipsMgDL()
-  await checkPatientListColumnSort()
-  await checkMonitoringAlertsIconsInactiveForFirstPatient()
-  await checkPatientListFilters()
-  await checkPatientColumnsFiltersContent()
-  await checkPatientListHideShowColumns()
-  await checkPatientListPendingTab(router)
-  await checkPatientListCurrentTab()
-}
-
-export const testPatientListForHcpWithMmolL = async () => {
-  await checkPatientListTooltipsMmolL()
-}
-
-export const testPatientListForHcpPrivateTeam = async () => {
-  await checkPatientListCurrentTabForPrivateTeam()
-  await checkPatientFiltersForPrivateTeam()
+export const testNotificationManagementForCaregiver = async () => {
+  await checkAcceptNotificationDirectShare()
 }